cleaning the test cases
